Move maybe_await to helpers.py
