+++ aioitertools/builtins.py
@@ -13,39 +13,21 @@ use with `await`, `async for`, etc.
 
 import asyncio
 import builtins
-import inspect
 from typing import (
     Any,
     AsyncIterable,
     AsyncIterator,
-    Awaitable,
     Callable,
     Iterable,
     List,
     Set,
     Tuple,
-    Union,
     cast,
     overload,
 )
 
+from .helpers import maybe_await
 from .types import AnyIterable, AnyIterator, AnyStop, T, R, T1, T2, T3, T4, T5
-
-
-@overload
-async def maybe_await(object: Awaitable[T]) -> T:
-    pass
-
-
-@overload
-async def maybe_await(object: T) -> T:
-    pass
-
-
-async def maybe_await(object: Union[Awaitable[T], T]) -> T:
-    if inspect.isawaitable(object):
-        return await object  # type: ignore
-    return object  # type: ignore
 
 
 def iter(itr: AnyIterable[T]) -> AsyncIterator[T]:
